requirements.txt
# Data Handling & Environment
python-dotenv
pandas
numpy
tqdm

# Jupyter Notebook
ipykernel

# Kaggle API
kaggle

# Image Processing & Computer Vision
opencv-python
pillow
albumentations

# Deep Learning (Core)
tensorflow

# Optional but recommended for Notebooks/Progress
matplotlib
scikit-learn